Fix bad GL version requests on NVPR-less systems
Requesting GLES 3.1 is pretty bad when the system only had <= 3.0.

Therefore, isSupported() has to gain support for functioning even when
there is no current context, and examples have to be changed to only
call format() when NVPR is reported to be available.

This removes the need for the previous patch's disabling of the compatibility
profile since we won't hit that path anymore with Intel.