Triangulate in QQuickItem::updatePolish()
Instead of doing it on the render thread in updatePaintNode(), perform
triangulating on the gui thread in updatePolish(). This involves breaking
the renderer interface up into gui and render thread specific functions,
mirroring QQuickItem's design. (object lives on gui thread, most functions
are only called on gui thread, updatePaintNode being the exception that is
invoked on the render thread with gui blocked)