Avoid triangulating when !visible
Apps may naively do visible: false for a subtree while continuing to animate.
With the default triangulating backend changing the geometry in any way leads
to retriangulating. Avoid this at least when we know for sure that the item
is not visible. Do it only when it becomes visible again.

This does not handle other cases, like opacity==0, but it better than nothing.